backend/requirements.txt
Django>=5.2
psycopg2-binary
djangorestframework
pandas
numpy
django-cors-headers
gunicorn 